0.032
Variavel de atk speed da reliquia

+++ click/Assets/Scripts/Reliquias/RelicData.cs
@@ -23,7 +23,7 @@ public class RelicData : ScriptableObject
     [Tooltip("Soma o valor a variavel de damage da torre")]
     public int towerAtkDamage;
     [Tooltip("Soma o valor a variavel de atkspeed da torre")]
-    public int towerAtkSpeed;
+    public float towerAtkSpeed;
     [Tooltip("Soma o valor a variavel de custo da torre")]
     public int custTower;
     [Tooltip("Soma o valor a variavel de drop de recurso")]
